Use floats instead of ints when unmarshalling JSON.
JSON, like JavaScript (naturally), has a single number type. Drivers may serialize numbers as floats, even if the numbers are intended to be integers. Lots of JSON libraries tack a ".0" on the ends of integers, so we should just be tolerant of float-y numbers.

PiperOrigin-RevId: 162693557

+++ go/launcher/webdriver/webdriver_test.go
@@ -411,8 +411,8 @@ func TestSetWindowSize(t *testing.T) {
 
 	testCases := []struct {
 		name   string
-		width  uint64
-		height uint64
+		width  float64
+		height float64
 		check  bool
 		err    bool
 	}{
@@ -486,8 +486,8 @@ func TestSetWindowPosition(t *testing.T) {
 
 	testCases := []struct {
 		name  string
-		x     int64
-		y     int64
+		x     float64
+		y     float64
 		check bool
 		err   bool
 	}{
